requirements.txt
-i https://pypi.org/simple
asgiref==3.8.0; python_version >= '3.8'
autopep8==2.1.0; python_version >= '3.8'
django==5.0.3; python_version >= '3.10'
django-filter==24.1; python_version >= '3.8'
djangorestframework==3.15.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
markdown==3.6; python_version >= '3.8'
pycodestyle==2.11.1; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2024.1; sys_platform == 'win32'
